# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CRUD accounts",
        "description": "CRUD accounts",
        "contact": {
            "name": "namnd"
        },
        "version": "1.0"
    },
    "paths": {
        "/api/accounts": {
            "get": {
                "tags": [
                    "Accounts"
                ],
                "summary": "Get a list of accounts",
                "operationId": "13efa5706d1fae14fc1cf87c9f6b1427",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of accounts per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of accounts",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Account"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Accounts"
                ],
                "summary": "Create a new account",
                "operationId": "e974666bc6c050296a8ee09a3ab6533f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AccountRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Account created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Account"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/accounts/{id}": {
            "get": {
                "tags": [
                    "Accounts"
                ],
                "summary": "Get account details by ID",
                "operationId": "552b29d7944dfff7a0c49abc83f44ea5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Account ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Account details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Account"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Account not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Accounts"
                ],
                "summary": "Update an account",
                "operationId": "9be8085be375b01725706b78df130bca",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Account ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AccountRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Account updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Account"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Account not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Accounts"
                ],
                "summary": "Delete an account",
                "operationId": "102c60c5fd1e94be8c71c4efbeb51bf8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Account ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Account deleted successfully"
                    },
                    "404": {
                        "description": "Account not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "AccountRequest": {
                "title": "Account Request",
                "description": "Request body for creating or updating an account",
                "required": [
                    "login",
                    "password",
                    "phone"
                ],
                "properties": {
                    "login": {
                        "description": "User login name, must be unique",
                        "type": "string",
                        "maxLength": 20
                    },
                    "password": {
                        "description": "User password",
                        "type": "string",
                        "format": "password",
                        "maxLength": 40,
                        "minLength": 6
                    },
                    "phone": {
                        "description": "User phone number",
                        "type": "string",
                        "maxLength": 20
                    }
                },
                "type": "object"
            },
            "Account": {
                "title": "Account",
                "description": "Account model",
                "properties": {
                    "id": {
                        "description": "Account ID",
                        "type": "integer"
                    },
                    "login": {
                        "description": "Login username",
                        "type": "string"
                    },
                    "password": {
                        "description": "Hashed password",
                        "type": "string"
                    },
                    "phone": {
                        "description": "Phone number",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Accounts",
            "description": "Accounts"
        }
    ]
}